Code now works for 2d and 3d Arrays
The original code uses as input a 3d array, i.e., two variables for the shape of the supra-adjacency matrix and one variable for the number of matrixes (individuals). If we have a single supra-adjacency matrix, we often represent it as a 2d array, which would yield a bug in our code.  Although the original code was not wrong, we decided to facilitate the user experience by adapting the code so that it also accepts a single supra-adjacency matrix as a 2d array as input. When that happens, the number of individuals in the analysis will be assigned one by default in the code.