The code also accepts 2d arrays as input
The original code uses as input a 3d array, i.e., two variables for the shape of the supra-adjacency matrix and one variable for the number of matrixes. If we have a single supra-adjacency matrix, we often represent it as a 2d array, which would yield a bug in our code.  Although the original code was not wrong, we decided to facilitate the user experience by adapting the code to accept a single supra-adjacency matrix as a 2d array as input. When that happens, the number of individuals in the analysis will be assigned one in the code.